Test kill all enemies objective and fix bugs

+++ Kthulhu-Wants-Me-Project/Assets/KthulhuWantsMe/Source/Gameplay/WaveSystem/WaveData.cs
@@ -8,7 +8,7 @@ namespace KthulhuWantsMe.Source.Gameplay.WaveSystem
     public class Batch
     {
         public List<EnemyPack> WaveEnemies;
-        public float NextBatchDelay;
+        public int NextBatchDelay;
     }
     [Serializable]
     public class WaveData
